Entity test
This is a super simple test in examples to feel out what entities might
look like. Since I don't want tables connected to entities, the entity
has to state how to collect the data it needs and then parse the result.
This simple way of doing it would cover 99% of usage, we just need to
add in a type-safe wrapper so that you don't have to call setup and
pass the results to the entity yourself. We can then look into adding
in things that are on Ktorm's EntitySequence.

The main problem with this right now is the inability to get results
based on the table name so if the query joins two tables with the same
column names, you can't fetch the duplicate columns with the column.
You have to switch to the index API which is confusing.