Remove sql_type
Trying to track down a bug where my app
is crashing when accessing an in list expression
on a binary expression when processing
a delete statement.

I found this issue
https://github.com/crystal-lang/crystal/issues/9667#issuecomment-710405767

I thought that the problem might be fixed
if I remove sql_type from my code
since there was a lot of usage of `T.class` or things
like it. But it didn't change anything.
I like the change so I'm keeping it